RHOAIENG-69608: Consolidate types-core.ts into @odh-dashboard/k8s-core package
