+++ frontend/src/__mocks__/mockDscStatus.ts
@@ -1,5 +1,4 @@
-import type { K8sCondition } from '@odh-dashboard/k8s-core';
-import { DataScienceClusterKindStatus } from '#~/k8sTypes';
+import type { K8sCondition, DataScienceClusterKindStatus } from '@odh-dashboard/k8s-core';
 import { DataScienceStackComponent } from '#~/concepts/areas/types';
 import { DataScienceStackComponentMap } from '#~/concepts/areas/const';
