+++ frontend/src/__mocks__/mockDsciStatus.ts
@@ -1,6 +1,8 @@
-import type { K8sCondition } from '@odh-dashboard/k8s-core';
+import type {
+  K8sCondition,
+  DataScienceClusterInitializationKindStatus,
+} from '@odh-dashboard/k8s-core';
 import { StackCapability } from '#~/concepts/areas/types';
-import { DataScienceClusterInitializationKindStatus } from '#~/k8sTypes';
 
 export type MockDsciStatus = {
   conditions?: K8sCondition[];
