+++ frontend/src/app/__tests__/AboutDialog.spec.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
 
-import type { DashboardConfigKind } from '@odh-dashboard/k8s-core';
+import type {
+  DashboardConfigKind,
+  DataScienceClusterInitializationKindStatus,
+  DataScienceClusterKindStatus,
+} from '@odh-dashboard/k8s-core';
 import { ClusterState, UserState } from '#~/redux/selectors/types';
 import { useUser, useClusterInfo } from '#~/redux/selectors';
 import { useAppContext } from '#~/app/AppContext';
@@ -10,11 +14,7 @@ import useFetchDsciStatus from '#~/concepts/areas/useFetchDsciStatus';
 import useFetchDscStatus from '#~/concepts/areas/useFetchDscStatus';
 import { mockDashboardConfig } from '#~/__mocks__';
 import { BuildStatus, SubscriptionStatusData } from '#~/types';
-import {
-  DataScienceClusterInitializationKindStatus,
-  DataScienceClusterKindStatus,
-  StorageClassKind,
-} from '#~/k8sTypes';
+import { StorageClassKind } from '#~/k8sTypes';
 import { FetchState } from '#~/utilities/useFetchState';
 import AboutDialog from '#~/app/AboutDialog';
 import { useWatchOperatorSubscriptionStatus } from '#~/utilities/useWatchOperatorSubscriptionStatus';
