+++ frontend/src/concepts/areas/useFetchDscStatus.ts
@@ -1,6 +1,6 @@
+import type { DataScienceClusterKindStatus } from '@odh-dashboard/k8s-core';
 import axios from '#~/utilities/axios';
 import useFetchState, { FetchState } from '#~/utilities/useFetchState';
-import { DataScienceClusterKindStatus } from '#~/k8sTypes';
 
 /**
  * Should only return `null` when on v1 Operator.
