+++ frontend/src/concepts/areas/useFetchDsciStatus.ts
@@ -1,6 +1,6 @@
+import type { DataScienceClusterInitializationKindStatus } from '@odh-dashboard/k8s-core';
 import axios from '#~/utilities/axios';
 import useFetchState, { FetchState } from '#~/utilities/useFetchState';
-import { DataScienceClusterInitializationKindStatus } from '#~/k8sTypes';
 
 /**
  * Should only return `null` when on v1 Operator.
