+++ packages/cypress/cypress/tests/mocked/modelServing/runtime/servingRuntimeList.cy.ts
@@ -17,11 +17,8 @@ import {
   mockInvalidTemplateK8sResource,
   mockServingRuntimeTemplateK8sResource,
 } from '@odh-dashboard/internal/__mocks__/mockServingRuntimeTemplateK8sResource';
-import type {
-  DataScienceClusterKindStatus,
-  InferenceServiceKind,
-  ServingRuntimeKind,
-} from '@odh-dashboard/internal/k8sTypes';
+import type { DataScienceClusterKindStatus } from '@odh-dashboard/k8s-core';
+import type { InferenceServiceKind, ServingRuntimeKind } from '@odh-dashboard/internal/k8sTypes';
 import { ServingRuntimePlatform } from '@odh-dashboard/internal/types';
 import { DataScienceStackComponent } from '@odh-dashboard/internal/concepts/areas/types';
 import { mockDsciStatus } from '@odh-dashboard/internal/__mocks__/mockDsciStatus';
