+++ packages/observability/src/utils/monitoringStackStatus.ts
@@ -1,5 +1,7 @@
-import type { K8sCondition } from '@odh-dashboard/k8s-core';
-import type { DataScienceClusterInitializationKindStatus } from '@odh-dashboard/internal/k8sTypes';
+import type {
+  K8sCondition,
+  DataScienceClusterInitializationKindStatus,
+} from '@odh-dashboard/k8s-core';
 
 export type MonitoringStatus =
   | { available: true }
